+++ main.js
@@ -8,7 +8,7 @@ import { doneButtonListners } from "./doneButtonListners.js";
 import { editButtonListners } from "./editButtonListners.js";
 import { enableButtonOnInput } from "./enableButtonOnInput.js";
 import { setInputValidation } from "./setInputValidation.js";
-import { massageSendButton } from "./api.js";
+import { massageSendButton } from "./massageSendListners.js";
 import { lastCommentDeleteButton } from "./lastCommentDelete.js";
